requirements.txt
langchain-text-splitters==0.3.8
langchain-chroma==0.2.4
langchain-community==0.3.24
pypdf==5.6.0
sentence-transformers==4.1.0
furhat-remote-api==1.0.2
python-dotenv==1.1.0